refactor: Transactional readOnly 추가
